refactor(cli): extract init/wire-conport into commands/bootstrap_commands.py
Move the 'init' and 'wire-conport' commands out of the cli.py monolith
into a dedicated module. cli.py now imports and registers them via
cli.add_command(). Path(__file__).parents adjusted for the new depth.

Test patches targeting dopemux.cli.init_project / ClaudeConfigurator
retargeted to dopemux.commands.bootstrap_commands — those symbols are
now looked up in the new module's namespace at call time.

Baseline CLI test suite still shows the 4 known pre-existing failures
(cli_group_help, save_command_{success,not_initialized,force_flag});
no new regressions.
